% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_volcano_plots_sc_metabolism.R
\name{display_individual_plots}
\alias{display_individual_plots}
\title{Display Individual Volcano Plots}
\usage{
display_individual_plots(
  individual_plots,
  cell_types = NULL,
  pause_between = FALSE
)
}
\arguments{
\item{individual_plots}{List of ggplot objects from create_individual_volcano_plots()}

\item{cell_types}{Character vector of cell types to display. If NULL, displays all. Default: NULL}

\item{pause_between}{Logical. Whether to pause between plots (interactive). Default: FALSE}
}
\description{
Convenience function to display individual volcano plots one by one
}
